refactor: Rename files and update import paths for consistency
