+++ src/utils/structs/auth/SignInRequestStruct.ts
