+++ src/utils/structs/auth/SignUpRequestStruct.ts
