+++ src/utils/structs/product/CreateProductRequestStruct.ts
