+++ src/utils/structs/product/UpdateProductRequestStruct.ts
