+++ src/utils/structs/user/GetMyFavoritesProductListRequestStruct.ts
