+++ src/utils/structs/user/GetMyProductListRequestStruct.ts
