+++ src/utils/structs/user/UpdatePasswordRequestStruct.ts
